+++ inst/shiny/server.R
@@ -1,37 +1,47 @@
 
 server <- function(input, output, session) {
 
-  # Startup message
-  output$scdf_summary <- renderPrint(cat(res$msg$startup))
+  # scdf ----
+  
+  ## startup message ----
+  
+  output$scdf_messages <- renderPrint(cat(res$msg$startup))
 
+  ## Render ----
   my_scdf <- reactiveVal()
-
   scdf_render <- reactive({
-
+    
     if (!inherits(my_scdf(), "scdf")) validate(res$msg$no_case)
-
-    output$scdf_summary <- renderPrint({
-      do.call("summary", list(my_scdf()))
-    })
-
-    output$scdf_syntax <- renderPrint({
-      req(inherits(my_scdf(), "scdf"))
-      do.call("convert", list(my_scdf(), inline = as.logical(input$convert)))
-    })
+    
+    if (input$scdf_output_format == "Summary"){
+      output$scdf_output <- renderPrint({
+        do.call("summary", list(my_scdf()))
+      })
+    } else if (input$scdf_output_format == "Syntax") {
+      output$scdf_output <- renderPrint({
+        req(inherits(my_scdf(), "scdf"))
+        do.call("convert", list(
+          my_scdf(), inline = as.logical(input$scdf_syntax_phase_structure)
+        ))
+      })
+    }
 
   })
 
+  observeEvent(input$scdf_output_format, scdf_render()) 
+  
+  ## input example ----
   observeEvent(input$scdf_example, {
     if (input$scdf_example != "(none)") {
       my_scdf(paste0("scan::", input$scdf_example) |> str2lang() |> eval())
       scdf_render()
+      output$scdf_messages <- renderPrint(cat(paste0("loaded example ", input$scdf_example)))
     } else {
       my_scdf(NULL)
     }
   })
 
-
-  # scdf: upload / save ------
+  ## upload (load) ------
   observeEvent(input$upload, {
     ext <- tools::file_ext(input$upload$datapath)
     if (ext == "rds") {
@@ -47,40 +57,50 @@ server <- function(input, output, session) {
     }
 
     if (!inherits(new, "scdf")) {
-      output$scdf_summary <- renderText(
-        "Sorry,\n the file you tried to upload is not a valid scdf file.")
+      output$scdf_messages <- renderText(
+        "Sorry,\n the file you tried to load is not a valid scdf file.")
     } else {
       my_scdf(new)
       scdf_render()
+      output$scdf_messages <- renderPrint(cat(paste0("loaded file successfully")))
     }
 
   })
 
+  ## download (save) ----
   output$scdf_save <- downloadHandler(
     filename = function() {
       scdf <- my_scdf()
       out <- paste(
-        input$prefix_output_data,
+        input$scdf_save_prefix,
         sprintf("%02d", length(scdf)),
         paste0(unique(scdf[[1]]$phase), collapse = ""),
         format(Sys.time(), format = "%y%m%d-%H%M%S"),
         sep = "-"
       )
-      paste0(out, input$save_scdf_format)
+      paste0(out, input$scdf_save_format)
     },
+    
     content = function(file) {
       scdf <- my_scdf()
-      if (input$save_scdf_format == ".rds") 
+      
+      if (!inherits(scdf, "scdf")) {
+        output$scdf_messages <- renderPrint(cat(res$error_msg$scdf_save))
+      } else {
+        output$scdf_messages <- renderPrint(cat("Saved file"))
+      }
+      
+      if (input$scdf_save_format == ".rds") 
         saveRDS(scdf, file)
-      if (input$save_scdf_format == ".R") 
+      if (input$scdf_save_format == ".R") 
         convert(scdf, file = file)
-      if (input$save_scdf_format == ".csv") 
+      if (input$scdf_save_format == ".csv") 
         write_scdf(scdf, filename = file)
+
     }
   )
 
-  # scdf: new cases --------
-
+  ## new cases --------
   observeEvent(input$add_case, {
     tryCatch({
       values <- paste0("c(", trim(input$values), ")")
@@ -114,6 +134,7 @@ server <- function(input, output, session) {
       if (input$remove_which == "last") {
         if (length(my_scdf()) > 0) new <- c(my_scdf(), new)
         my_scdf(new)
+        output$scdf_messages <- renderPrint(cat("Appended case"))
         scdf_render()
       } 
       
@@ -128,19 +149,20 @@ server <- function(input, output, session) {
             new <- c(my_scdf()[1:(at-1)], new, my_scdf()[at:(length(my_scdf()))])
           }
           my_scdf(new)
+          output$scdf_messages <- renderPrint(cat("Added case at position", input$remove_at))
           scdf_render() 
         }
       }  
 
     },
     error = function(e)
-      output$scdf_summary <- renderText(
+      output$scdf_messages <- renderText(
         paste0(res$error_msg$invalid_case, "\n\n", e)
       )
     )
   })
 
-  # scdf: remove cases --------
+  ## remove cases --------
   observeEvent(input$remove_case, {
     if (input$remove_which == "last") {
       if (length(my_scdf()) > 1) {
@@ -154,17 +176,21 @@ server <- function(input, output, session) {
         my_scdf(my_scdf()[-input$remove_at])
     }
     
+    output$scdf_messages <- renderPrint(cat("removed case"))
     scdf_render()
   })
 
-  # scdf: remove all cases --------
+  ## remove all cases --------
   observeEvent(input$remove_all, {
     my_scdf(NULL)
+    output$scdf_messages <- renderPrint(cat("Cleared cases"))
     scdf_render()
+    
   })
 
-  # transform ----
+  # Transform ----
 
+  ## render ----
   transformed <- reactive({
     out <- my_scdf()
     syntax = "scdf"
@@ -226,31 +252,30 @@ server <- function(input, output, session) {
     out
   })
 
-  # transform save ----
-  
+  ## save ----
   output$transformed_save <- downloadHandler(
     filename = function() {
       scdf <- transformed()
       out <- paste(
-        input$prefix_output_transformed,
+        input$transform_save_prefix,
         sprintf("%02d", length(scdf)),
         paste0(unique(scdf[[1]]$phase), collapse = ""),
         format(Sys.time(), format = "%y%m%d-%H%M%S"),
         sep = "-"
       )
-      paste0(out, input$save_transformed_format)
+      paste0(out, input$transform_save_format)
     },
     content = function(file) {
-      if (input$save_transformed_format == ".rds") 
+      if (input$transform_save_format == ".rds") 
         saveRDS(transformed(), file)
-      if (input$save_transformed_format == ".R") 
+      if (input$transform_save_format == ".R") 
         convert(transformed(), file = file)
-      if (input$save_transformed_format == ".csv") 
+      if (input$transform_save_format == ".csv") 
         write_scdf(transformed(), filename = file)
     }
   )
   
-  # transform output ------
+  ## output ------
   
   output$transform_scdf <- renderPrint({
     if(!inherits(my_scdf(), "scdf")) validate(res$msg$no_case)
@@ -262,8 +287,9 @@ server <- function(input, output, session) {
     export(transformed(), caption = "") |> HTML()
   })
   
-  # stats -----
+  # Stats -----
 
+  ## Calculate ---- 
   calculate_stats <- reactive({
     if (!inherits(my_scdf(), "scdf")) validate(res$msg$no_case)
     scdf <- transformed()
@@ -275,6 +301,7 @@ server <- function(input, output, session) {
     )
   })
 
+  ## Output ----
   output$stats_html <- renderUI({
     results <- calculate_stats()
     print_args <- input$stats_print_arguments
@@ -306,8 +333,7 @@ server <- function(input, output, session) {
   })
 
 
-  # stats: arguments ------
-
+  ## Arguments ------
   stat_arg_names <- reactive({
     args <- names(formals(input$func))
     values <- formals(input$func)
@@ -399,7 +425,7 @@ server <- function(input, output, session) {
     call
   })
 
-  # stats save ------
+  ## Save ------
   
   output$stats_save <- downloadHandler(
     
@@ -447,8 +473,9 @@ server <- function(input, output, session) {
     }
   )
   
-  # plot -----
+  # Plot -----
 
+  ## Render ----
   render_plot <- reactive({
     req(inherits(my_scdf(), "scdf"))
     call <- paste0("scplot(transformed())")
@@ -497,10 +524,25 @@ server <- function(input, output, session) {
 
   observeEvent(input$plot_arguments, render_plot_syntax())
 
-  output$plot_scdf <- renderPlot({
+  ## Output ----
+ 
+  render_plot_syntax <- reactive({
+    call <- paste0("scplot(scdf)")
+    if (trimws(input$plot_arguments) != "") {
+      call <- paste0(
+        call, "%>%\n  ", gsub("\n", " %>%\n  ", trimws(input$plot_arguments))
+      )
+    }
+    output$plot_syntax <- renderPrint({
+      cat(call)
+    })
+  })
+  
+  output$plot_scdf <- renderPlot(res = 120,{
     render_plot()
   })
 
+  ## Save ----
   output$saveplot <- downloadHandler(
     filename = function() {
       scdf <- transformed()
@@ -521,18 +563,99 @@ server <- function(input, output, session) {
     }
   )
 
-  render_plot_syntax <- reactive({
-    call <- paste0("scplot(scdf)")
-    if (trimws(input$plot_arguments) != "") {
-      call <- paste0(
-        call, "%>%\n  ", gsub("\n", " %>%\n  ", trimws(input$plot_arguments))
-      )
-    }
-    output$plot_syntax <- renderPrint({
-      cat(call)
-    })
-  })
+  # Power test -----
+  
+  output$pt_results <- renderPrint(cat(res$placeholder$pt))
+  
+  ## Render -----
+  render_power_test <- reactive({
+    syntax <- paste0(
+      "design(\n  n = ", input$design_n, ", ",
+      "phase_design = list(", input$design_phase, "), \n  ",
+      "trend = ", input$design_trend, ", ",
+      "level = list(", input$design_level, "), ",
+      "slope = list(", input$design_slope, "), \n  ",
+      "start_value = ", input$design_start, ", ",
+      #"s = ", input$design_s, ", ",
+      "rtt = ", input$design_rtt, ", ",
+      "distribution = ", deparse(input$design_distribution),
+      "\n)"
+    )  
+    
+    ci <- input$pt_ci
 
+    syntax <- paste0(
+      syntax, " %>% \n",
+      "  power_test(method = ", deparse(input$pt_method), ", ",
+      "effect = ", deparse(input$pt_effect), ", ",
+      "n_sim = ", input$pt_n, ", ",
+      "ci = ", ci,
+      ")"
+    )
+    
+    syntax
+  })
+  
+  ## Output ----
+  output$pt_syntax <- renderPrint({
+    cat(render_power_test())
+  })
+  
+  ## Plot ----
+  
+  observeEvent(input$desigh_plot_button, {
+    call <- paste0(
+      "design(\n  n = ", input$design_n, ", ",
+      "phase_design = list(", input$design_phase, "), \n  ",
+      "trend = ", input$design_trend, ", ",
+      "level = list(", input$design_level, "), ",
+      "slope = list(", input$design_slope, "), \n  ",
+      "start_value = ", input$design_start, ", ",
+      #"s = ", input$design_s, ", ",
+      "rtt = ", input$design_rtt, ", ",
+      "distribution = ", deparse(input$design_distribution),
+      "\n) %>% ",
+      "random_scdf() %>% ",
+      "scplot()"
+    )
+    
+    output$plot_design <- renderPlot(res = 100, {
+      str2lang(call) |> eval()
+    })
+    #tryCatch(
+    #  str2lang(call) |> eval(),
+    #  error = function(x)
+    #    output <- renderText(paste0(res$error_msg$plot, "\n\n", x))
+    #)
+
+  })
+  
+  ## Analyse ----
+  
+  observeEvent(input$pt_compute, {
+    
+    phase_structure <- eval(str2lang(
+      paste0("list(", input$design_phase, ")")
+    ))
+    if (length(phase_structure) > 2) {
+      output$pt_results <- renderPrint({
+        cat("Sorry, power-tests are only possible for designs with two phases.")
+      })
+    } else {
+      call <- render_power_test()
+      output$pt_results <- renderPrint({cat("Calculating ...")})
+      res <- tryCatch(
+        str2lang(call) |> eval(),
+        error = function(e)
+          paste0("Sorry, could not proceed calculation:\n\n", e)
+      )
+      output$pt_results <- renderPrint({
+        if (inherits(res, "character")) cat(res) else res
+      })
+    }
+    
+  })
+  
   # quit app -----
   observeEvent(input$navpage, {
     if (input$navpage == "Quit") {